updating crawler thread for new memory management
page and robots mem_mgrs have to be instantiated with a configuration object.
This has to be done within the main thread method, however constructing and
passing configuration objects does not look clean.

Config objects could be created within the crawler thread object constructor Im
not sure that this is a clean solution